+++ davraw9705/convertrawfile.py
@@ -1,8 +1,13 @@
 import datetime as dt
 import os
+import time
 
 import numpy as np
 import pandas as pd
+
+import files
+import plots
+import stats
 
 pd.set_option('display.width', 1000)
 pd.set_option('display.max_columns', 15)
@@ -44,37 +49,55 @@ class ConvertRawFile():
     SCALING_FACTOR_co2 = 1.114
     SCALING_FACTOR_h2o = 1.114
 
-    def __init__(self, raw_file, indir, outdir, log, raw_file_fullpath, aux_file_fullpath):
+    def __init__(self, raw_file, indir, outdir_rawdata_ascii, log, raw_file_fullpath, aux_file_fullpath, run_id,
+                 numfiles, filecounter, outdir_plots_hires, profile_factors_df):
         self.indir = indir
-        self.outdir = outdir
+        self.outdir = outdir_rawdata_ascii
         self.log = log
         self.raw_file = raw_file
         self.raw_file_fullpath = raw_file_fullpath
         self.aux_file_fullpath = aux_file_fullpath
+        self.run_id = run_id
+        self.numfiles = numfiles
+        self.filecounter = filecounter
+        self.outdir_plots_hires = outdir_plots_hires
+        self.profile_factors_df = profile_factors_df
 
+        self.filestats_df = pd.DataFrame()
         self.set_colnames()  # Define column names in raw file
 
-        self.raw_filename_start_time = self.get_file_start_time(ext='raw')
+        self.raw_filename_start_time = \
+            self.get_file_start_time(ext='raw',
+                                     filename_parsing_string=self.filename_parsing_string,
+                                     raw_file=self.raw_file)
+
+        self.raw_filename_date = pd.to_datetime(self.raw_filename_start_time.date())
 
         self.run()
 
-    def get_file_start_time(self, ext):
-        _parser = f"{self.filename_parsing_string}.{ext}"
-        return dt.datetime.strptime(self.raw_file, _parser)
+    @staticmethod
+    def get_file_start_time(ext, filename_parsing_string, raw_file):
+        """Get file start time from filename"""
+        _parser = f"{filename_parsing_string}.{ext}"
+        return dt.datetime.strptime(raw_file, _parser)
 
     def run(self):
-        self.log.info(f"")
-        self.log.info(f"{self.raw_file}")
-        self.log.info("=" * 20)
+        tic = time.time()  # Measure processing time for file
+
+        self.log.info("")
+        perc = (self.filecounter / self.numfiles) * 100
+        self.log.info(f"{self.raw_file}  (File #{self.filecounter} of {self.numfiles})  ({perc:.1f}%)")
+
+        self.log.info("=" * 50)
         self.log.info(f"    raw path:  {self.raw_file_fullpath}")
         self.log.info(f"    aux path:  {self.aux_file_fullpath}")
         self.log.info(f"    start time from filename:  {self.raw_filename_start_time}")
 
-        # .raw file
+        # Get data from .raw file
         RAW_df, filled_date_range, rawfile_starttime_str = \
             self.get_raw_file_data()
 
-        # .aux file
+        # Get data from .aux file
         AUX_df = \
             self.get_aux_file_data(filled_date_range=filled_date_range)
 
@@ -84,21 +107,88 @@ class ConvertRawFile():
                                aux_df=AUX_df)
 
         # Corrections
-        MERGED_df = self.apply_calibration_coefficients(df=MERGED_df)
+        MERGED_df = self.apply_calibration_coefficients(df=MERGED_df.copy())
+        # MERGED_df = self.apply_corrections(df=MERGED_df.copy())  # todo act
 
         # Make multi-row column index from tuples
         MERGED_df.columns = pd.MultiIndex.from_tuples(MERGED_df.columns)
 
+        # data_start_dt = MERGED_df.index[0]
+        # data_end_dt = MERGED_df.index[-1]
+        # data_expected_end_dt = self.raw_filename_start_time + pd.Timedelta(hours=1)
+        # if data_end_dt > data_expected_end_dt:
+        # file_duration = MERGED_df.index[0] - MERGED_df.index[-1]
+        # self.raw_filename_start_time
+
+        # hh_groups = MERGED_df.groupby(pd.Grouper(freq='30T'))  # Half-hourly groups in hourly file
+        # for name, group in hh_groups:
+        #     print(group.index[0])
+            # group.to_csv(string.replace(str(name), ':', '_') + '.csv')
+
+        # # Plot high-res data  # todo act
+        # plots.high_res_ts(df=MERGED_df, outfile=self.raw_file, outdir=self.outdir_plots_hires, logger=self.log)
+
+        # Stats collection
+        self.filestats_df = stats.calc(df=MERGED_df.copy(), bin_filedate=self.raw_filename_start_time, logger=self.log)
+
+        # Save data
         self.save_file(df=MERGED_df, rawfile_starttime_str=rawfile_starttime_str)
 
-    def set_colnames(self):
+        time_needed = time.time() - tic
+        self.log.info(f"    (Time needed: {time_needed:.2f}s)")
 
+    def apply_corrections(self, df):
+        """Correct measurements based on comparison IRGA vs profile"""
+
+        co2_factor = self.profile_factors_df.loc[self.raw_filename_date]['co2_factor']
+        h2o_factor = self.profile_factors_df.loc[self.raw_filename_date]['h2o_factor']
+
+        self.log.info(f"    Applying correction factors from IRGA vs profile comparison "
+                      f"to CO2 and H2O measurements: CO2 * {co2_factor}  |  H2O * {h2o_factor} ...")
+
+        co2_corrected = df[self.co2_col].multiply(co2_factor)
+        h2o_corrected = df[self.h2o_col].multiply(h2o_factor)
+
+        co2_corrected.name = self.co2_col
+        h2o_corrected.name = self.h2o_col
+
+        df.update(co2_corrected)
+        df.update(h2o_corrected)
+
+        return df
+
+    def get_filestats(self):
+        return self.filestats_df
+
+    def set_colnames(self):
+        """
+        Set column names for converted files
+
+        Column names already include the correct units after conversion.
+        Note that the original raw data files contain column names that
+        show wrong information, e.g. CO2 is recorded as mV signal but
+        later converted to µmol/mol.
+        """
+
+        # Raw data columns
         self.u_col = ('U', '[m/s]')
         self.v_col = ('V', '[m/s]')
         self.w_col = ('W', '[m/s]')
-        self.sos_col = ('T_SONIC', '[degC]')
-        self.co2_col = ('C', '[µmol/mol]')
-        self.h2o_col = ('Xw', '[mmol/mol]')
+        self.t_sonic_col = ('T_SONIC', '[degC]')  # In degC after conversions
+
+        # Conversion of CO2 and H2O signals is tricky
+        # The units given in the header are wrong.
+        # Given values in the .raw files are voltages in mV.
+        # Conversions are documented in the LI-6262 manual.
+        # Below the final units for CO2 and H2O after conversions in this script are given.
+
+        # CO2 concentration, in umol/mol after conversion, according to manual.
+        # I assume this is dry mole fraction, i.e. mixing ratio
+        self.co2_col = ('CO2', '[µmol/mol]')
+
+        # H2O vapor concentration, in mmol/mol after conversion, according to manual
+        self.h2o_col = ('H2O', '[mmol/mol]')
+
         self.pa_col = ('Pa', '[kPa]')  # Air pressure
         self.plic_col = ('Plic', '[kPa]')  # IRGA pressure (Licor)
         self.ta_col = ('Ta', '[degC]')  # Air temperature
@@ -106,6 +196,7 @@ class ConvertRawFile():
         self.xa_col = ('Xa', '[mmol / mol]')
         self.rhoa_col = ('Rhoa', '[kg / m3]')
 
+        # Additional columns
         self.timestamp_col = ('TIMESTAMP', '[yyyy-mm-dd HH:MM:SS.]')
         self.index_col = ('Index', '[-]')
         self.date_col = ('Date', '[-]')
@@ -126,10 +217,13 @@ class ConvertRawFile():
         """Read .raw files and prepare data"""
 
         # Read raw file
-        RAW_df = self.read_file(filepath=self.raw_file_fullpath, header=[3, 4])
+        RAW_df = self.read_file(filepath=self.raw_file_fullpath,
+                                header_rows_list=[0, 1],  # Rows AFTER skip rows considered
+                                skip_rows_list=[0, 1, 2],
+                                header_section_rows_list=[0, 1, 2, 3, 4])
 
-        # Add empty units and rename T
-        RAW_df = self.sanititze_raw_colnames(df=RAW_df)
+        # Add empty units
+        RAW_df = self.add_empty_units(df=RAW_df)
 
         # Remove unnamed columns
         RAW_df = self.remove_unnamed_cols(df=RAW_df)
@@ -137,6 +231,9 @@ class ConvertRawFile():
         # Make timestamp
         RAW_df, filled_date_range, rawfile_starttime_str = \
             self.create_raw_timestamp(df=RAW_df)
+
+        # Rename colnames
+        RAW_df = self.sanitize_raw_colnames(df=RAW_df)
 
         # Convert to units
         RAW_df = self.convert_raw_to_units(df=RAW_df)
@@ -146,10 +243,13 @@ class ConvertRawFile():
     def get_aux_file_data(self, filled_date_range):
         """Find respective .aux file, read it and prepare data"""
 
-        AUX_df = self.read_file(filepath=self.aux_file_fullpath, header=[1, 2])  # read file
+        AUX_df = self.read_file(filepath=self.aux_file_fullpath,
+                                header_rows_list=[0, 1],  # Rows AFTER skip rows considered
+                                skip_rows_list=[0],
+                                header_section_rows_list=[0, 1, 2])
 
         # Add empty units, rename T and remove unnamed columns
-        AUX_df = self.sanititze_raw_colnames(df=AUX_df)
+        AUX_df = self.add_empty_units(df=AUX_df)
         AUX_df = self.remove_unnamed_cols(df=AUX_df)
 
         # Make timestamp
@@ -159,34 +259,50 @@ class ConvertRawFile():
                                                       keep_timestamp_col=True)  # make timestamp
         return AUX_df
 
-    def sanititze_raw_colnames(self, df):
-        # Rename units for columns where units were empty
+    def sanitize_raw_colnames(self, df):
+        """Rename raw data columns"""
+
+        # aux_cols = [('Pa', '[kPa]'), ('Plic', '[kPa]'), ('Ta', '[degC]'),
+        #             ('Tlic', '[degC]'), ('Xa', '[mmol/mol]'), ('Rhoa', '[kg/m3]')]
+
+        # Original column names and renamed column names
+        # For T the wrong original units are in the original header file
+        # (original units are sonic temperature x 50).
+        orig_colnames_dict = {
+            ('U', '[m/s]'): self.u_col,
+            ('V', '[m/s]'): self.v_col,
+            ('W', '[m/s]'): self.w_col,
+            ('Sound', '[m/50s]'): self.t_sonic_col,
+            ('T', '[degC]'): self.t_sonic_col,
+            ('C', '[mV]'): self.co2_col,
+            ('C', '[µmol/mol]'): self.co2_col,
+            ('Xw', '[mV]'): self.h2o_col,
+            ('Xw', '[mmol/mol]'): self.h2o_col
+        }
+
+        # Rename columns
+        new_colnames = []
+        for orig_col in df.columns:
+            if orig_col in orig_colnames_dict:
+                new_col = orig_colnames_dict[orig_col]
+                if new_col != orig_col:
+                    self.log.info(f"    Renamed column {orig_col} to {new_col}")
+            else:
+                new_col = orig_col
+                self.log.info(f"    (!)WARNING: Column {new_col} is not a defined raw data column.")
+            new_colnames.append(new_col)
+
+        df.columns = new_colnames
+        return df
+
+    def add_empty_units(self, df):
+        """Rename units for columns where units were empty"""
         new_colnames = []
         for col in df.columns:
             varname = col[0]
             varunits = col[1]
-
             if "Unnamed:" in varunits:
                 varunits = "[-]"
-
-            # For T the wrong original units are in the original header file (original units are
-            # sonic temperature x 50)
-            # Here, T is renamed and the units are kept the same
-            if (varname == 'T') & (varunits == '[degC]'):
-                varname = "T_SONIC"
-                varunits = "[degC]"
-
-            # # PUT UNITS IN HEADER
-            # # units = df.iloc[0]
-            # df = df.reindex(df.index.drop(0))
-            # for idx, col in enumerate(df.columns):
-            #     if str(units[idx]) != 'nan':
-            #         if (col == 'T') and (
-            #                 units[idx] == '[degC]'):  # for T the wrong units are in the original header file
-            #             df = df.rename(columns={col: 'SOS_[m/s]'})
-            #         else:
-            #             df = df.rename(columns={col: col + '_' + str(units[idx])})
-
             new_colnames.append((varname, varunits))
         df.columns = new_colnames
         return df
@@ -196,23 +312,107 @@ class ConvertRawFile():
         Identify unnamed columns
 
         Some years have too many tabs in the file, more tabs than data columns
-        empty tabs are read with col name 'Unnamed*' we use this info to delete
+        empty tabs are read with col name 'Unnamed*' (or 'unkown') we use this info to delete
         empty data columns, so our data file only contains actual data values.
         Remove columns with empty (unnamed) variable name
         """
         for col in df.columns:
             varname = col[0]
             if "Unnamed:" in varname:
-                df = df.drop(col, 1)  # todo check if works
+                df = df.drop(col, 1)
                 self.log.info(f"    (!) Removed unnamed column {col}")
+            elif "unknown" in varname:
+                df = df.drop(col, 1)
+                self.log.info(f"    (!) Removed unknown column {col}")
         return df
 
-    def save_file(self, df, rawfile_starttime_str):
+    def save_file(self, df: pd.DataFrame, rawfile_starttime_str: str):
         # SAVE FILE
-        new_filename = os.path.join(self.outdir, 'CH-DAV_' + rawfile_starttime_str + '.csv')
-        df.to_csv(new_filename, index=True)
-        # contents.to_csv(new_filename, index=False)
-        self.log.info("    --> saved to file {}".format(new_filename))
+        new_filename = os.path.join(self.outdir, 'CH-DAV_' + rawfile_starttime_str + '.csv.gz')
+        # df.to_csv(new_filename, index=True)
+        df.to_csv(new_filename, index=False, compression='gzip')
+        self.log.info(f"    --> saved merged RAW/AUX data file {new_filename}")
+
+    # def apply_h2o_corrections(self, df):
+    #     """
+    #
+    #     Use the parameters attached to correct our measured H2O concentrations (H2O_EC)
+    #     for up to August 2005. I used a simple linear regression with Nabel_H2O to get
+    #     these parameters.
+    #
+    #     H2O.corrected = slope * H2O_EC + intercept
+    #
+    #     TIMESTAMP.START	         TIMESTAMP.END          slope       intercept   Adjusted R-squared
+    #     1997-01-01 00:30:00 	 1998-01-01 00:00:00 	1.087378	24.757392	0.9723
+    #     1998-01-01 00:30:00 	 1999-01-01 00:00:00 	1.171352	1.241446	0.9994
+    #     1999-01-01 00:30:00 	 2000-01-01 00:00:00 	1.1655515	2.7512004	0.9893
+    #     2000-01-01 00:30:00 	 2001-01-01 00:00:00 	1.1617209	3.9202374	0.9977
+    #     2001-01-01 00:30:00 	 2002-01-01 00:00:00 	1.1473303	4.2208861	0.9978
+    #     2002-01-01 00:30:00 	 2003-01-01 00:00:00 	1.1644168	5.9501428	0.9945
+    #     2003-01-01 00:30:00 	 2004-01-01 00:00:00 	1.1580541	1.0469319	0.9953
+    #     2004-01-01 00:30:00 	 2005-01-01 00:00:00 	1.304316	-12.997544	0.9485
+    #     2005-01-01 00:30:00 	 2005-08-09 16:30:00 	1.2226907	-3.389643	0.9976
+    #
+    #     """
+    #
+    #     if pd.to_datetime('1997-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('1998-01-01 00:00:00'):
+    #         slope = 1.087378
+    #         intercept = 24.757392
+    #
+    #     elif pd.to_datetime('1998-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('1999-01-01 00:00:00'):
+    #         slope = 1.171352
+    #         intercept = 1.241446
+    #
+    #     elif pd.to_datetime('1999-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('2000-01-01 00:00:00'):
+    #         slope = 1.1655515
+    #         intercept = 2.7512004
+    #
+    #     elif pd.to_datetime('2000-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('2001-01-01 00:00:00'):
+    #         slope = 1.1617209
+    #         intercept = 3.9202374
+    #
+    #     elif pd.to_datetime('2001-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('2002-01-01 00:00:00'):
+    #         slope = 1.1473303
+    #         intercept = 4.2208861
+    #
+    #     elif pd.to_datetime('2002-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('2003-01-01 00:00:00'):
+    #         slope = 1.1644168
+    #         intercept = 5.9501428
+    #
+    #     elif pd.to_datetime('2003-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('2004-01-01 00:00:00'):
+    #         slope = 1.1580541
+    #         intercept = 1.0469319
+    #
+    #     elif pd.to_datetime('2004-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('2005-01-01 00:00:00'):
+    #         slope = 1.304316
+    #         intercept = -12.997544
+    #
+    #     elif pd.to_datetime('2005-01-01 00:30:00') <= self.raw_filename_start_time <= \
+    #             pd.to_datetime('2005-08-09 16:30:00'):
+    #         slope = 1.2226907
+    #         intercept = -3.389643
+    #
+    #     else:
+    #         slope = 1
+    #         intercept = 0
+    #
+    #     h2o_measured = df[self.h2o_col].copy()
+    #     h2o_corrected = h2o_measured.multiply(slope).add(intercept)
+    #     self.log.info(f"    Corrected H2O values H2O_corrected = {slope} * H2O + {intercept}")
+    #
+    #     # Update data in df with calibrated data
+    #     h2o_corrected.name = self.h2o_col  # Set name to enable update of data
+    #     df.update(h2o_corrected)
+    #
+    #     return df
 
     def apply_calibration_coefficients(self, df):
         df.replace('None', np.nan, inplace=True)
@@ -331,7 +531,6 @@ class ConvertRawFile():
         starttime_str = first_datetime.strftime("%Y%m%d%H%M")
         _readable_starttime_str = first_datetime.strftime("%Y-%m-%d %H:%M:%S")
 
-
         df.drop([self.date_col, self.time_col, self.year_col, self.month_col, self.day_col,
                  self.hour_col, self.minute_col, self.second_col, 'temp'], axis=1, inplace=True)
         if keep_timestamp_col:
@@ -343,7 +542,11 @@ class ConvertRawFile():
         # generate continuous date range and re-index data
         # filled_date_range = pd.date_range(df.index[0], df.index[-1], freq=this_file_freq)
         df = df.sort_index()
-        df = df.reindex(filled_date_range, fill_value=-9999, method='nearest')  # apply new continuous index to data
+        df = df.reindex(filled_date_range, fill_value=-9999, method='nearest')  # Old method
+        # df = df.reindex(filled_date_range)  # apply new continuous index to data
+        # df = df.interpolate(method='linear', axis=0, limit=None, inplace=False)  # Linear interpolation for missing vals
+        # df = df.fillna(method='backfill', inplace=False)  # Back-fill to get values for start of file
+        # df = df.fillna(method='ffill', inplace=False)  # Forward-fill to make sure there are no missing vals
 
         # df to numeric
         df = df.apply(pd.to_numeric, args=('coerce',))
@@ -355,8 +558,6 @@ class ConvertRawFile():
         # df.replace(np.nan, -9999, inplace=True)  # should not be necessary
 
         return df
-
-
 
     def get_multipliers(self):
         """the h2o and co2 had some amplifier installed at certain times"""
@@ -385,11 +586,20 @@ class ConvertRawFile():
 
         # Wind speeds u, v, w
         uw_multiplier = 0.01
-        v_multiplier = -0.01  # Note: we mirror the y-axis of the R2A
+
+        # Note: in the CDEF files, generated by using the conversion script 'wsl2cdef', the
+        # y-axis of the R2A is mirrored (multiplied by -1).
+        # To my understanding in the CDEF files the wind direction is not defined according
+        # to the instrument north arrow, but along the u wind speed axis. However, when using
+        # raw data in EddyPro mirroring is not necessary. The sonic is oriented with 270° North
+        # offset in relation to the North arrow. We can use this info directly in EddyPro for
+        # calculations of the correct wind direction. Wind directions look correct with these
+        # settings (270°, Axis).
+        v_multiplier = 0.01
 
         # Speed of sound
         # Solent R2/R2A sonics multiply the SOS (measured in m s-1) w/ a factor of 50
-        # and then report them as a positive integer value todo
+        # and then report them as a positive integer value
         sos_multiplier = 0.02
 
         return uw_multiplier, v_multiplier, sos_multiplier, co2_multiplier, h2o_multiplier
@@ -435,7 +645,7 @@ class ConvertRawFile():
         df[self.timestamp_col] = self.raw_filename_start_time
 
         # Create total runtime in seconds
-        frequency = 0.0480192076830732  # seconds; 20.825 Hz = 1 value every 0.0480192076830732 seconds
+        frequency = 0.048000000768  # seconds; 20.833333 Hz = 1 value every 0.048000000768 seconds
         df[self.total_runtime_col] = df.index
         df[self.total_runtime_col] = df[self.total_runtime_col].multiply(frequency)
 
@@ -464,31 +674,47 @@ class ConvertRawFile():
 
         return df, filled_date_range, starttime_str
 
-    def read_file(self, filepath, header):
+    def read_file(self, filepath, skip_rows_list: list, header_rows_list: list,
+                  header_section_rows_list: list):
         try:
-            df = pd.read_csv(filepath, header=header, delimiter='\t', engine='python')
+            # df = pd.read_csv(filepath, header=header, delimiter='\t', engine='python', index_col=None)
             # df = pd.read_csv(file_fullpath, header=header, delimiter='\t', engine='python', tupleize_cols=True)
-        except:
-            try:
-                # Some files (very few) have "Error: line contains NULL byte"
-                # this means there is NUL in the file when looked at it in a text editor
-                # with option "Show All Characters" or similar seems there is way to
-                # solve this following the code from:
-                #   http://osdir.com/ml/python-pydata/2012-05/msg00044.html:
-                fi = open(filepath, 'r')
-                data = fi.read()
-                fi.close()
-                fo = open(filepath + '_NULLBYTES-REMOVED', 'w')
-                fo.write(data.replace('\x00', ''))  # remove NULL BYTES
-                fo.close()
 
-                df = pd.read_csv(filepath + '_NULLBYTES-REMOVED', header=header, delimiter='\t',
-                                 engine='python')
-                # df = pd.read_csv(file_fullpath + '_NULLBYTES-REMOVED', header=header, delimiter='\t', engine='python', tupleize_cols=True)
-            except:
-                df = pd.read_csv(filepath + '_NULLBYTES-REMOVED', header=header, delimiter='\t',
-                                 error_bad_lines=False)
-                # df = pd.read_csv(file_fullpath + '_NULLBYTES-REMOVED', header=header, delimiter='\t', tupleize_cols=True, error_bad_lines=False)
+            df, _ = files.parse_csv_file(filepath=filepath,
+                                         skip_rows_list=skip_rows_list,
+                                         header_rows_list=header_rows_list,
+                                         header_section_rows_list=header_section_rows_list)
+
+
+        except Exception as e:
+            self.log.info(f"    (!)WARNING Exception caught in file {filepath}")
+            self.log.info(f"    (!)WARNING     Error message: \"{e}\"")
+            self.log.info(f"    (!)WARNING     Now removing NULLBYTES from file ...")
+
+            # Some files (very few) have "Error: line contains NULL byte"
+            # this means there is NUL in the file when looked at it in a text editor
+            # with option "Show All Characters" or similar seems there is way to
+            # solve this following the code from:
+            #   http://osdir.com/ml/python-pydata/2012-05/msg00044.html:
+            fi = open(filepath, 'r')
+            data = fi.read()
+            fi.close()
+            _file_nullbytes = f"{filepath}_NULLBYTES-REMOVED.aux"
+            fo = open(_file_nullbytes, 'w')
+            fo.write(data.replace('\x00', ''))  # remove NULL BYTES
+            fo.close()
+
+            self.log.info(f"    (!)WARNING     Done.")
+            self.log.info(f"    (!)WARNING     Trying to parse NULLBYTES-REMOVED ...")
+
+            try:
+                df = pd.read_csv(_file_nullbytes, header=header_rows_list, delimiter='\t', engine='python')
+                self.log.info(f"    (!)WARNING     Worked.")
+            except Exception as e:
+                self.log.info(f"    (!)WARNING     Failed!")
+                self.log.info(f"    (!)WARNING     Trying to parse NULLBYTES-REMOVED file by skipping error lines ...")
+                df = pd.read_csv(_file_nullbytes, header=header_rows_list, delimiter='\t', error_bad_lines=False)
+                self.log.info(f"    (!)WARNING     Worked.")
 
         self.log.info(f"    {filepath.name}  columns:  {len(df.columns)}  /  rows:  {len(df)}")
         return df